Optionally override Scanner base path
This is principally for testing purposes, because in actual use the
intended purpose of the add-on is to always operate in the current
working directory.

In testing, however, it is useful to provide a helper method (in the
spec_helper.rb file) to generate a new Scanner object configured to
search the "fixtures" directory, without having to explicitly change
directories within the specs themselves.

Signed-off-by: Wincent Colaiuta <win@wincent.com>