Remove matches_for method from CommandT::Base
CommandT::Base is really just a convenience class for providing public
access to the other classes in the library which do the real work of
matching and scanning.

As such, direct access to the low-level "matches_for" method really
isn't required, as all public users of the class will be more interested
in the "sorted_matches_for" method.

Signed-off-by: Wincent Colaiuta <win@wincent.com>