Call clearmatches() on creating matches window
This clears any custom highlighting rules that might have been added to
the window automatically when it was created using VimEnter autocmds
(for example, I use this to apply special highlighting to overlength
lines).

In this way we get full control over how lines are highlighting in the
matches listing. Note the effects are not permament (do not apply to
other windows).

Signed-off-by: Wincent Colaiuta <win@wincent.com>