Save and restore hlsearch setting
If a users has hlsearch set to true and happens to search for a string
which is part of a file name, the string will be highlighted when we
bring up the matches list.

So set nohlsearch on showing the list, and restore whatever the old
value of hlsearch was on closing.

Signed-off-by: Wincent Colaiuta <win@wincent.com>