Refactoring of focus tracking
In order to implement type ahead find we need to juggle things around a
little bit.

Both the MatchWindow and the Prompt need to know when they have focus,
but it should be the responsibility of the Controller to inform them of
when this is the case. The Controller should only send key press
messages to one of the other objects, and only when it is appropriate to
do so (for example, the delete key isn't sent at all when the
MatchWindow has the focus).

Signed-off-by: Wincent Colaiuta <win@wincent.com>