Special handling or equalalways setting
Turn it off before creating the match window, so that VIM won't waste
time trying to auto-adjust the size of all the other visible windows.

Then, after deleting the match window, restore the previously existing
window sizes and then turn equalalways back on (if the user had it on).
In this way such users will get their desired behaviour (of
auto-adjusting splits).

The behaviour should now be very close to correct for all cases; for
example, when a user asks to open a new file in a split window, we
close the match window, restore the existing sizes, reselect the
previously selected window, then perform the split. This should
appropriately use the available space and dump the cursor in the new
window.

Signed-off-by: Wincent Colaiuta <win@wincent.com>