Restore window sizes in tallest-to-shortest order
When restoring window sizes we have to do it in a particular order to
handle cases involving vertical splits like this:

  ----------------
     1    |
  --------|   2
     3    |
  ----------------

When collapsed to make room for the match window these windows might
look like:

  ----------------
  --------|   2
  ----------------

Afterwards when we try to restore them all to their original size the
ordering is very important.

For example, if we try to restore window 1 first we will fail because
the height of window 2 serves as a constraint and there is not enough
room.

To avoid this we make sure that we always restore the size of the taller
windows first. In this case that means we resize window 2 first, so that
the resizes of windows 1 and 3 will succeed.

For more information see:

  https://wincent.com/issues/1493

Signed-off-by: Wincent Colaiuta <win@wincent.com>