Remove ! from function declarations
As the plugin uses a "load guard" variable to prevent itself from being
evaluated multiple times, there is no real need for these "!" on
function declarations.

The "!" would allow us to silently redefine a pre-existing function, but
given that we won't be evaluated multiple times and our function names
are all "protected" by the CommandT "namespace", we really shouldn't
ever need to silently redefine.

In the case where we actually do end up redefining something it is
almost certainly by accident, so we should probably allow the error to
bubble up where the user might notice it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>