Update specs for dot-file changes
Change paths specifed in "./foo/bar" format to "foo/bar".

This brings the specs into line with actual usage in VIM, where paths
are always specified relative to the current working directory (without
the leading "./").

It also fixes some breakage in the specs caused by the recent addition
of dot-file filtering, which treats "./" as a dot-file seeing as it is a
path-component starting with a dot.

Signed-off-by: Wincent Colaiuta <win@wincent.com>