Treat "list" setting as window-local rather than global
The "list" setting was being incorrectly treated as a global setting
rather than a window-local one. This meant that whatever the value for
the current window at the time Command-T was invoked ended up being
saved and then restored globally afterwards, possibly causing an
incorrect "list" setting to be used in other windows.

See:

  https://wincent.com/issues/1515

Signed-off-by: Wincent Colaiuta <win@wincent.com>