Suppress warning when Cursor highlight group not found
I was of the understanding that the Cursor highlight group would exist
in all color schemes seeing as the VIM docs (:help hl-Cursor) mention it
as one of the "default highlighting groups"; neverthless, it appears
that it is possible for it not to exist, at least according to user
reports:

  https://wincent.com/issues/1513

It's not clear whether this happens when syntax highlighting isn't
enabled, or only with some colorschemes, or only on some terminals.
Regardless of the cause, make sure that any "E411: highlight group not
found: Cursor" errors which might occur don't bubble up to the annoy
the user, and to be doubly cautious, don't even try looking at the
Cursor highlight if syntax highlighting isn't even enable.

Signed-off-by: Wincent Colaiuta <win@wincent.com>