Remove multiruby functionality from Rakefile
The approach used in the Rakefile was never going to work for running
the specs in conjunction with a C Ruby extension, because the extension
would need to be rebuilt between each run of "spec spec" on different
versions.

As there is no practical way of doing this, the simpler solution is just
to create a shell script which does the builds and runs "spec spec"
after each one. This was added in the last commit, so the stuff in the
Rakefile is officially redundant now.

Signed-off-by: Wincent Colaiuta <win@wincent.com>