Add "g:CommandTMatchWindowAtTop" setting
The best match is usually what I'm interested in, but as I'm narrowing
down the results, its location on my screen keeps shifting downward
(since it appears at the top of a bottom-anchored window). This commit
provides a g:CommandTMatchWindowAtTop boolean option, defaulting to false,
which makes the match window appear at the top of the screen, thus
keeping the best match visible at the top line.

Signed-off-by: Wincent Colaiuta <win@wincent.com>