Remove explicit check that options param is a hash
Previously we had an explicit check that the supplied options parameter
was a hash, raising an exception with the message "options not a hash".

This check is unnecessary because:

  - the only user of the classes involved (Match and Matcher) is the
    Controller class itself, and we can guarantee that the correct
    option type is always passed
  - the classes involved are not public API so we don't have to worry
    about the input being "untrustworthy"
  - even if somebody does improperly access the API and pass in a
    non-hash options parameter, an exception will be raised naturally
    anyway when we try to send a message:

      NoMethodError: undefined method `has_key?'

This change might also fix the following issue where the "options not a
hash" exception was being raised at runtime, but not in the IRB console,
on one user's Ruby 1.9.1 install. I am not sure of the cause (perhaps
some problem with header files), but removing the check can only bring
us one step closer to solving the problem:

  https://wincent.com/issues/1520

Signed-off-by: Wincent Colaiuta <win@wincent.com>