Don't map <Esc> for some terminal types
As vt100 and xterm terminals send the arrow keys as sequences like
<Esc>OA, VIM gets confused by bindings for the <Esc> key for those
terminals (it actually interprets the arrow keys as sequences of
commands).

For now we avoid the problem on those terminals by not mapping <Esc> at
all, but the long-term solution will be to provide a means for users to
override the default mappings.

See:

  https://wincent.com/issues/1531

Signed-off-by: Wincent Colaiuta <win@wincent.com>