Use hash to specify mappings
This is a refactoring in preparation for a subsequent commit in which
we will provide a means for users to override the default mappings. The
hash structure allows us to conveniently process a bunch of different
mappings -- some of which are unique and others which are multiple --
and issue the call to the "map" method in a single place.

Signed-off-by: Wincent Colaiuta <win@wincent.com>