Add "trouble-shooting" label to help and reference it
The Stub object will only be initiated if the load path is correctly set
up, so if we get that far we can basically assume that the problem lies
somewhere with the C extension.

So instead of complaining about not being able to load "the necessary
modules", explicitly mention the C extension and point people at the
relevant section on the documentation.

Signed-off-by: Wincent Colaiuta <win@wincent.com>