Handle non-existent directories passed to :CommandT
Gracefully handle non-existent directories passed in via ":CommandT
path".

Note that we catch the Errno:ENOENT exception only when the match window
is shown, because we're interested in advising the user when they enter
a dud path. (In contrast, we let Errno::ENOENT exceptions at other times
just bubble up, as may be the case if the user has something funky with
their filesystem and stuff appears in the listing but later claims to
not exist.)

This closes feature request #1522:

  https://wincent.com/issues/1522

Signed-off-by: Wincent Colaiuta <win@wincent.com>