Use more conservative language standard setting
Previously we used "gnu99" as the C language standard. At least one
user, however, has reported about getting warnings with this setting and
having to manually edit the Makefile to set it to "c99":

  http://amix.dk/blog/post/19501

So, seeing as we don't use any GNU extensions, try switching the std
setting to "c99" and see if that makes any difference for such users.

Signed-off-by: Wincent Colaiuta <win@wincent.com>