Bring wildignore behavior closer to VIM's own behaviour
In commit f314c1e6 we tried to make Command-T react to the 'wildignore'
setting in way that more closely matches VIM's own behaviour by
using the builtin expand() function which takes into account
'wildignore'.

While this works quite well, there are still some discrepancies from
VIM's actual behaviour.

For example, while patterns like this work as expected with no
surprises:

  *.o
  .git
  **/build

A pattern like this:

  vendor/rails/**

Will work differently in Command-T than it does in the rest of VIM.
Specifically, if I type this in VIM:

  :e vendor/rails/<tab>

Then VIM's autocomplete won't allow me to drill down into the directory
because it is excluded by the 'wildignore' setting.

Command-T, on the other hand, will allow me to see the contents of that
directory. This is because internally, at the time the file names are
checked, they are almost always absolute paths, because the default
starting directory is ":pwd" which is itself provided by VIM as an
absolute path. As such, VIM's expand() function checks to see if:

  /absolute/path/to/vendor/rails/

Matches against the 'wildignore' pattern of:

  vendor/rails/**

And decides that it does not match.

As a work around, a user could specify a pattern like this:

  **/vendor/rails/**

But it is a bit ugly because it doesn't fit well with the behaviour of
VIM itself.

The fix, then, is to not pass absolute paths into the expand() function,
but to pass paths relative to the starting directory.

In 99% of cases, the starting directory is the ":pwd", so the behaviour
should then be identical to VIM's own behaviour.

In the cases where the user has passed in an override for the starting
directory (either relative or absolute), then the behaviour will diverge
slightly from VIM's while still hopefully being consistent and intuitive
from the user's perspective.

For example, if :pwd is the HOME directory and the user invokes
Command-T with:

  :CommandT path/to/some/rails/project

Then a 'wildignore' which includes this pattern:

  vendor/rails/**

Will exclude all files under:

  path/to/some/rails/prokect/vendor/rails/

Which is probably what the user expects. For comparison, typing:

  :e path/to/some/rails/project/vendor/rails/<tab>

Will autocomplete despite the 'wildignore' settings, which is consistent
with the standard VIM behaviour as described previously.

Signed-off-by: Wincent Colaiuta <win@wincent.com>