Handle "wildcard" characters in filenames
The recent change (commit f314c1e) to get VIM to handle exclusion
checking for us via the expand() function will choke on paths with
"special" wildcard characters in them (see 'h: wildcards' for a
list).

For example, a path like:

  /foo/bar/notes [Foo Support #abcdef-058505].txt

Will cause en E16 range error when fed into expand() due to the "[" and
"]" characters which are interpreted as delimiting an invalid range.

So we escape:

  ?
  * (and by extension, **)
  [
  ]
  `

Unfortunately, the docs say "Which wildcards are supported depends on
the system", and lists "the common ones". We can only hope that the
above list covers most cases and will have to handle any breakage which
arises on specific platforms in the future.

Note also, that in this commit we switch form expand() to glob(),
because glob() takes a smaller range of "special" characters (just he
wildcards) and so is less likely to choke on paths with special
characters in them.

Signed-off-by: Wincent Colaiuta <win@wincent.com>