Mock VIM calls to get all specs passing again
Recent commits (starting with f314c1e) changed the design of the Scanner
class, such that rather than performing filename globbing internally for
the purposes of exclusion, it "shelled out" back to VIM itself so that
it could use VIM's "expand()" function and thus handle a wider range of
exclusion patterns in the 'wildignore' setting.

This, of course, broke the Scanner specs, and the Finder specs too,
because the specs were always intended to be run in isolation from VIM,
and in fact there _is_ no straightfoward way of running them in any
other fashion.

So here we use mocking and stubbing to get the specs passing again when
running in isolation outside of the context of VIM. Later on will be
adding "integration" tests which can be run from inside VIM itself to
make sure that things actually work as expected.

Signed-off-by: Wincent Colaiuta <win@wincent.com>