Add base functions for obtaining "best score" via recursion
As noted in this ticket:

  https://wincent.com/issues/1598

Search results might be sorted in a more intuitive order if instead of
using a naive left-to-right scoring algorithm, we consider other
possible ways of matching and pick the one that yields the best score.

The basic algorithm is this: after finding a matching letter, bump the
"cursor" one character to the right and re-try matching via recursion to
see if a better score is produced.

In this way we scan all possible matchings and the one with the best
score ends up "bubbling up" out of the recursion.

It remains to be seen whether this will have acceptable performance, but
I want to try it out.

Before this can actually be used I have to insert the actual scoring
algorithm; this is just the code required to implement the recursion
itself.

Signed-off-by: Wincent Colaiuta <win@wincent.com>