Remove garbage collection book-keeping
After the refactoring, we no longer have to worry about a Ruby exception
being raised after allocating storage for the matchinfo_t struct. The
only method which might conceivably raise is the
CommandT_option_from_hash function, and that is called before
allocating the struct.

In fact, in a subsequent commit we can just place the struct on the
stack and avoid allocation entirely.

Signed-off-by: Wincent Colaiuta <win@wincent.com>