Switch to RSpec 2, RR, Bundler
I've been seeing suspicious segfaults during spec suite runs and had
noticed that they mostly occurred when calling the "paths" method on
a scanner instance.

As this particular call is mocked or stubbed, I wondered if it might be
fixed by changing to another mocking framework. That is what this commit
does.

The easiest way I know to do this and have full control over the
versions involved is to use Bundler, so I'm switching to that in this
commit too.

Note that it doesn't actually fix the segfaults. Further investigation
shows that it has nothing to do with mocking/stubbing; the common
trigger is literally _any_ use of rb_funcall in the extension at all.

I suspect some mismatch between the version of Ruby used in the spec
environment and the version used at runtime from inside Vim, because the
latter works perfectly.

So, even though this switch doesn't fix the problem it set out to, I am
going to keep it as it brings us a nicer mocking/stubbing syntax.

Signed-off-by: Wincent Colaiuta <win@wincent.com>