Tweak max-score-per-char calculation
Previously early matches always won. ie. given files:

  doc/command-t.txt
  TODO

A search for "d" would always rank the first path more highly (with a
score of 1.0, actually, which was never really intended). This was
because the max score per char was calculated as 1 / abbrev_len.

We recently changed that to 1 / str_len. This was necessary to ensure
that given two matching paths and all else being equal, the shortest
path should win; eg. searching for "art":

  articles.rb
  articles_controller.rb

The first path should win over the former. In practice things often
turned out this way because both of these would score a perfect score of
1.0 under the old algorithm, and then alphabetical sorting would take
care of the rest. Under the new algorithm we guarantee that the shorter
match wins.

This however, produced unexpected results in the case of the initial
example; now we have the scores for "d" inverted:

  TODO
  doc/command-t.txt

In fact, even searching for "do" maintains that ordering.

Evidently some tweaks are required. This commit gets at least the "d"
case tuned to where we want it, although the "do" case still remains to
be done.

Let the two methods of determing max score per character be called
"method 1" and "method 2".

Method 1: as the number of characters in the search string increases
their importance diminishes

Method 2: the number of characters in the search string is irrelevant
and each character is always worth the same; it is the length of the
path which causes max score per character to change

So, in reality, paths never change length, they are fixed inputs, but we
will see the number of chars in the search string growing.

We take the average of "method 1" and "method 2" and use that as the max
score per char. In this way at small character counts, method 1 is
dominant -- this solves our unintuitive result for "d" -- but as the
user types more characters method 2 soon becomes more dominant. I
believe this will give us more intuitive results at longer search string
lengths.

As I said, the "do" problem remains to be solved.

Signed-off-by: Wincent Colaiuta <win@wincent.com>