Try to avoid "Not enough room" errors
This commit implements an attempt to avoid "Not enough Room" errors
which can be reported if we try opening a file when a small buffer like
MiniBufExplorer is active.

Specifically we target "unlisted" buffers of buftype "nofile". Note that
we merely try to switch to the first window containing a buffer which is
not "unlisted"/"nofile". There are no guarantees that said window will
be appropriate for us either, but it is still more likely to be
appropriate than the "unlisted"/"nofile" buffer.

Signed-off-by: Wincent Colaiuta <win@wincent.com>