Watch out for nil windows when restoring dimensions
While it is unlikely to occur, it is possible that VIM::Window[w.index]
may return nil for an existing window.

For example, starting with a freshly opened Vim window, if we open the
MiniBufExplorer window with "<leader>mbe" and then try to open a file
with Command-T, we'll find that one of the two windows (the original "No
Name" window and the MiniBufExplorer window) is no longer available and
VIM::Window[w.index] returns nil.

This problem does not manifest itself when calling up Command-T directly
from within the original "No Name" buffer (ie. without switching to
MiniBufExplorer first), nor when we have other, real buffers present and
we call up MiniBufExplorer.

Signed-off-by: Wincent Colaiuta <win@wincent.com>