Better fix for issue #1617
Even though I can't reproduce the Vim "largefile" bug, I did some
testing by stubbing $curbuf.number and forcing it to always return 0. It
turns out that my fallback in this case wasn't working because if you
do:

  :bunload! GoToFile

Vim will fail to close the window because the GoToFile buffer is hidden
and it can't find it.

  :bwipeout! GoToFile

However, does work, so go with that in the fallback case.

Signed-off-by: Wincent Colaiuta <win@wincent.com>