Tighten up check for existing mappings
Commit 70a8f9c9 added a check for an existing "CommandT" mapping before
setting up the default mapping, to allow for the case where a user has
set up a custom mapping in their .vimrc.

This, however, breaks when somebody adds a mapping for ":CommandTFlush",
as described here:

  https://wincent.com/issues/1762

This is because the Vim hasmapto() function does a relaxed match when
looking at the right-hand side of the mapping, so a mapping for
":CommandTFlush" is considered to match "CommandT".

Fix this by tightening up the check, looking for ":CommandT<CR>" instead
of just "CommandT".

Signed-off-by: Wincent Colaiuta <win@wincent.com>