Use vimball.rb to produce Vimball
One less thing to set up when developing on a different machine: instead
of having to find and build or install the mkvimball executable, we can
use the vendored vimball.rb script to do build the Vimball.

The world would be a better place if everyone used Pathogen, but if the
Vimball creation process is automated then there's no reason not to
continue doing it for now.

Signed-off-by: Wincent Colaiuta <win@wincent.com>