Rakefile: 'stash pop' failing is not necessarily a bad thing
If there are no local changes in the tree then the initial 'git stash'
will be a no-op, emitting:

  No local changes to save

Subsequently, when we do 'git stash pop' we'll exit with a non-zero exit
status and will see:

  No stash found.

As this is not really an error, don't make the entire Rake task explode
when this happens. If there really is an actual error at this point, the
user will see it from the message.

Signed-off-by: Wincent Colaiuta <win@wincent.com>