Move cursor to selected line in match window
In GUI Vim (eg MacVim) the cursor correctly hides, but when running in a
terminal it does not. By moving the cursor to the currently selected
line we at least minimize the visual distraction when it starts
blinking (at least it draws the eye to the selected line rather than
blinking in the upper-left corner).

For related discussion see:

  https://github.com/pMongoose/Command-T/commit/7e3e5f96b0bc5a2dfe4fd891c4ee0f77eb752332

It may be possible to teach some kinds of terminal, such as xterm, to
actually hide the cursor, but that will be a task for later development:

  http://vim.wikia.com/wiki/Configuring_the_cursor

Signed-off-by: Wincent Colaiuta <win@wincent.com>