Don't needlessly redisplay list when showing reversed lists
The "matches != @matches" comparison would spuriously pass for reversed
lists, except for the case where the list contains only 1 match, because
we were comparing the unreversed list with the reversed list.

So, reverse the list before doing the comparison, thus avoiding
unnecessary reprinting of the matches when the match list is unchanged.

Signed-off-by: Wincent Colaiuta <win@wincent.com>