Warn about making non-release gems
Abbreviating the full version number of an intermediate build (eg.
"1.1b2-10-g61a374a") to the last tag (eg. "1.1b2") is evil, because it
means we end up labelling distinct versions of the software with the
same version number.

Unfortuantely, however, it is a necessary evil because RubyGems will
choke if passed an intermediate version number:

  Invalid gemspec in [command-t.gemspec]: Malformed version number string
  1.1b2-10-g61a374a
  ERROR:  While executing gem ... (NoMethodError)
      undefined method `mark_version' for nil:NilClass
      rake aborted!

So, we resort to using the abbreviated, misleading version number in
this case. We can at least mitigate this evilness by displaying a
warning if the user ever tries to build a gem for an intermediate
release, by adding a dependency on the ":check_tag" task.

Signed-off-by: Wincent Colaiuta <win@wincent.com>