Remove "install" Rake task as it is causing confusion
Looks like some people/scripts saw the "rake install" task and thought
that was the way to install the Command-T plug-in in Vim, but it is not:

  https://github.com/carlhuda/janus/issues/163

Remove the task to avoid such confusion. People who want to install the
Gem can still do so with "gem install command-t", which is easy enough.

(And to anyone reading this commit message, note that the gem was added
so that third-parties could access the Command-T matching algorithm from
outside of Vim, but to get the full functionality from inside Vim, you
must continue to install it as always, as a Vim plug-in and not as a
gem. See the docs for how to do this.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>