Reverting lazy-initialization code, pending investigation
This commit rolls back the recently-added lazy-initialization code due
to a report of a segfault; due to the fact that a lot of people use
Pathogen or similar to track the HEAD of master, I'd rather be
conservative about this and roll it back immediately, and then
investigate. The lazy-initialization code itself has been moved onto a
topic branch (lazy-initialization).

The bug report:

  https://wincent.com/issues/1837

Signed-off-by: Wincent Colaiuta <win@wincent.com>