Teach CommandT::Stub to handle show_buffer_finder
If for some reason you didn't compile the extension before trying to use
the buffer finder you would trigger an error because the CommandT::Stub
class didn't implement the #show_buffer_finder method.

As the number of stubbed methods has now reached three, this commit also
includes a refactoring to eliminate the repetition in the method
definitions.

Signed-off-by: Wincent Colaiuta <win@wincent.com>