Use multiple path caches, keyed off directories
This is an initial, naïve implementation. All it does is store separate
caches for each path it is asked to scan. While this works fine, before
it can be released it needs a one key thing: people should be able to
turn it off, or otherwise configure how many different directories
should be cached.

Note that there's no cleverness here; if you work for a while in
directory "/foo" and later change into "/foo/bar", it scans the new
directory rather than trying to reuse the path info that it obtained
when scanning the first directory.

Requested in:
  https://wincent.com/issues/1878

Signed-off-by: Wincent Colaiuta <win@wincent.com>