Attempted fix for "E90: Cannot unload last buffer" error
Details noted in the code comment.

Note that in the case where the `:quit` command kicks in, the subsequent
`:bwipeout!`/`:bunload!` commands will still run, and most likely fail;
this seems harmless enough and visible error messages are suppressed
with `silent!`, but it seems sensible to err on the side of caution and
make sure clean-up really happens (the Vim docs are silent on what
`:quit` does to unlisted buffers, but experimentation with help windows,
at least, shows that they linger on in the `:ls!` output after a
`:quit`, even if their contents end up getting unloaded).

See:

  https://wincent.com/issues/1778

Signed-off-by: Wincent Colaiuta <win@wincent.com>