Add ability to move matches into quickfix window
Sometimes I want to edit a bunch of related files. For example, in the
Twigg project there are 5 gemspec files which I often have to edit
together. In this use case, it's convenient to be able to open
Command-T, type "gemspec" (or similar) and then move all the matches
into the quickfix window (from whence I can skip from file to file with
`:cn`, `:cnf` etc).

This commit makes this possible with a new mapping, defaulting to
`<C-q>` (mnemonic: "[q]uickfix") and overiddable using the
g:CommandTQuickfixMap setting, which takes the current match listing and
transfers it to the quickfix window. Additionally, it closes the match
listing and opens the quickfix window.

Signed-off-by: Wincent Colaiuta <win@wincent.com>