Switch to needle/haystack names for clarity
"str" and "abbrev" are not as clear as "haystack" and "needle",
respectively, so let's switch to the latter.

I'm doing this just inside the match.c functions, as this is where the
potential for confusion is most ripe. "abbrev" is still used in some other
places, such as in matcher.c, but there things seem less likely to
confuse, so I am leaving that as-is for now.

Signed-off-by: Wincent Colaiuta <win@wincent.com>