Short-circuit if needle can't possibly fit inside haystack
This is a cheap little optimization that prevents us form scanning (or
further recursing) if we're too close to the end of the haystack to
possibly encounter the rest of the needle. As it only applies towards
the end of the string, it's not going to make a huge impact, but every
little bit helps.

Signed-off-by: Wincent Colaiuta <win@wincent.com>