Add benchmarks
Here are the baseline results after recent optimizations:

                          user     system      total        real
  pathological        2.940000   0.000000   2.940000 (  2.958157)
  command-t           3.620000   0.030000   3.650000 (  3.652865)
  chromium (subset)   3.410000   0.020000   3.430000 (  3.437180)
  chromium (whole)   10.160000   0.050000  10.210000 ( 10.229278)

Short-circuiting for needles that can't possibly fit is a pretty big win
for some cases. The "pathological" case was basically untenable prior to
adding memoization (ie. it wouldn't finish in any reasonable timeframe),
although the memoization does appear to slow down the other cases by
about 3-4%.

The goal from here on will be to make the standard cases as fast and
faster than before, while still preserving the ability to handle the
pathological cases without grinding to a halt.

Signed-off-by: Wincent Colaiuta <win@wincent.com>