Switch to The One True Brace Style in C files
If you look carefully at the diff, you'll see that there were some
places where I was unthinkingly gravitating towards that style anyway,
so let's just give in and make the change.

The original brace style was most likely largely due to the influence of
Objective-C on me, of which I've written a fair bit in the last decade.

The new style is largely the result of my work at Causes, where we
happen to use The One True Brace Style in our JavaScript; the force of
habit is starting to make itself felt.

In addition to modifying the brace style, I tweaked a fair few lines for
readability, making a few things align and adding blank lines where I
felt it led to a better vertical rhythm. There are a couple of gnarly
and hard-to-read conditionals that could use some more work, but I
mostly just let Vim format those as it saw fit.

Signed-off-by: Wincent Colaiuta <win@wincent.com>