Combine match & matcher benchmarks
I'm about to start another round of optimizations in which the current
Match class will go away, becoming an internal implementation detail of
the Matcher class.

As such, the benchmark script needs to be adapted to only interact with
the latter.

This commit makes the necessary adjustments, folding the match
benchmarks into the matcher benchmarks. The overall results continue to
be basically the same.

Signed-off-by: Wincent Colaiuta <win@wincent.com>