Tweak Matcher specs to use OpenStructs instead of stubbing
I'm about to fold the Match specs into the Matcher class (or rather,
I'll be exercising the Match functionality via the Matcher specs, as the
Match class will soon become an internal-only implementation detail of
the Matcher class).

So, this commit tidies up the specs a little by switching to the use of
OpenStructs rather than stubbing. This will enable me to add a bunch of
different specs more easily, as no separate stubbing step will be
required.

Signed-off-by: Wincent Colaiuta <win@wincent.com>