Make options hash to Matcher#sorted_matches_for truly optional
There was a lying comment in there suggesting that the options parameter
hash was optional, when in fact it was not.

Noticed while writing specs for the method.

In the name of making life as easy as possible for callers, this commit
makes the comment about optionality true.

Included for free is a (cosmetic) typo fix in a nearby code comment.

Signed-off-by: Wincent Colaiuta <win@wincent.com>