Enumerate system headers exhaustively in extconf.rb
Let's be explicit about all the different headers we require (although
in practice, the additional headers listed here are much more "standard"
and likely to be present on a system, compared to "ruby.h").

When we add pthread support later on, we'll be adding an additional
check here for "pthread.h", which is another less "standard" header (at
least, on Windows).

Signed-off-by: Wincent Colaiuta <win@wincent.com>