Simplify Match instance initialization
This is a step towards turning the Match class into something that is
pure C and has no Ruby-specific calls in it, which is itself a necessary
prelude to the threaded performance work that I plan on doing later.
(Note, I suspect the reduction of Ruby object allocations that will be a
subsequent step will make this faster even before we get to the
threading bit.)

In a sense, this makes the initialization API strictly worse, replacing
an options hash with two explicit but superficially confusing option
names. Nevertheless, I'm not worried about that right now as the whole
file is going to be reworked as I get rid of the "Ruby-ness" of it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>