Add basic prototype of threaded searching
The code is pretty ugly, so I'll be looking at cleaning it up in
subsequent commits, as well as adding some configurability (automatic or
otherwise), but nevertheless it gives us a very nice speedup on large
codebases.

Before:
                          user     system      total        real
  pathological        0.000000   0.000000   0.000000 (  0.004329)
  command-t           0.420000   0.000000   0.420000 (  0.416208)
  chromium (subset)   1.070000   0.000000   1.070000 (  1.071452)
  chromium (whole)    3.680000   0.010000   3.690000 (  3.683207)

After:
                          user     system      total        real
  pathological        0.010000   0.030000   0.040000 (  0.018285)
  command-t           0.980000   1.850000   2.830000 (  1.391637)
  chromium (subset)   1.480000   0.020000   1.500000 (  0.551915)
  chromium (whole)    5.150000   0.020000   5.170000 (  1.861232)

Note that the pathological case is actually slightly slower now, but I
suspect that is because of the overhead of invoking a thread which isn't
really needed for the small number of paths (ie. one!) in sample data
set.

The slowdown of the command-t sample data is a little more alarming
though, so I might end up adding a threshold such that only if the
number of paths is greater than, say, 1,000, do we invoke the threading
machinery.

As written in this commit, the threaded support will only be compiled in
if you define USE_THREADS at build time. I'll be making this somewhat
more automatic in a future commit. I've defaulted to it being
effectively inactive so that this commit could, at least in theory, be
merged to the master branch even though I haven't figured out what to do
about Windows support yet.

As Ruby threads are basically useless (even in 1.9, due to the GIL), all
of this is done at a lower-level, using pthreads and avoiding any
interaction with the Ruby runtime (beyond making read-only access of
primitive Ruby objects already in memory). Once the work is done, all
the threads are joined and Ruby and Vim are free to go back to their
usual single-threaded operation.

The division of labor here is fairly crude: given four worker threads,
thread 0 will process paths 0, 4, 8 etc, and thread 1 will process paths
1, 5, 9 etc, and so on. I did it this what instead of dividing the paths
into contiguous chunks because I wanted to increase the likelihood of
each worker getting a similar amount of work (due to the way file
scanners work -- ie. recursively, using a depth-first search -- it is
possible or even likely that the distribution of the search space will
be uneven across the paths array (ie. shorter paths near the beginning
and end, with longer paths in the middle). By "striping" the work across
workers, we make it more likely that they'll all have roughly the same
about of work to do.

Another reason for this simple/crude division of labor is it frees us
from the overhead of having to manage a queue data structure, or a
producer/consumer set-up. Note we don't need to use any locks here as we
effectively have a map-reduce job that slices an array into 4 chunks,
and each thread only ever tries to mutate its own chunk. It is not even
necessary to zero-out the array, as the threads will take care of
filling it out.

Signed-off-by: Wincent Colaiuta <win@wincent.com>